Refactor update in board, user

+++ src/main/java/com/example/jpaboard/controller/BoardController.java
@@ -2,9 +2,12 @@ package com.example.jpaboard.controller;
 
 import com.example.jpaboard.dto.BoardDTO;
 import com.example.jpaboard.dto.CategoryDTO;
+import com.example.jpaboard.entity.User;
 import com.example.jpaboard.service.BoardService;
 import com.example.jpaboard.service.CategoryService;
+import com.example.jpaboard.service.UserService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -18,6 +21,7 @@ public class BoardController {
 
     private final CategoryService categoryService;
     private final BoardService boardService;
+    private final UserService userService;
 
     @ModelAttribute("categories")
     public List<CategoryDTO> categories() {
@@ -26,14 +30,26 @@ public class BoardController {
         return categories;
     }
 
+    @ModelAttribute("loginId")
+    public String loginId(Authentication auth) {
+        if (auth != null) {
+            return auth.getName();
+        }
+        return null;
+    }
+
     @GetMapping("/new")
-    public String createForm(Model model) {
+    public String createForm(Model model, Authentication auth) {
+        if (auth == null) {
+            return "redirect:/user/login";
+        }
+
         return "board/saveBoard";
     }
 
     @PostMapping("/new")
-    public String create(@ModelAttribute BoardDTO boardDTO) {
-        boardService.save(boardDTO);
+    public String create(@ModelAttribute BoardDTO boardDTO, Authentication auth) {
+        boardService.save(boardDTO, auth.getName());
 
         return "redirect:/";
     }
@@ -48,24 +64,38 @@ public class BoardController {
     }
 
     @GetMapping("/update/{id}")
-    public String updateForm(@PathVariable Long id, Model model) {
+    public String updateForm(@PathVariable Long id, Model model, Authentication auth) {
         BoardDTO board = boardService.findById(id);
+
+        if (!board.getUser().getLoginId().equals(auth.getName())) {
+            throw new IllegalStateException("잘못된 접근입니다.");
+        }
+
         model.addAttribute("board", board);
 
         return "board/updateBoard";
     }
 
-    @PostMapping("/update")
-    public String update(@ModelAttribute BoardDTO boardDTO, Model model) {
-        System.out.println(boardDTO);
-        BoardDTO updateDTO = boardService.update(boardDTO);
-        model.addAttribute("board", updateDTO);
+    @PostMapping("/{boardId}/update")
+    public String update(@PathVariable Long boardId, @ModelAttribute BoardDTO boardDTO, Model model) {
+        Long updateBoardId = boardService.update(
+                boardId, boardDTO.getCategory(), boardDTO.getTitle(), boardDTO.getContent());
+
+        BoardDTO updateBoard = boardService.findById(updateBoardId);
+
+        model.addAttribute("board", updateBoard);
 
         return "board/detailBoard";
     }
 
     @GetMapping("/delete/{id}")
-    public String delete(@PathVariable Long id) {
+    public String delete(@PathVariable Long id, Authentication auth) {
+        BoardDTO board = boardService.findById(id);
+
+        if (!board.getUser().getLoginId().equals(auth.getName())) {
+            throw new IllegalStateException("잘못된 접근입니다.");
+        }
+
         boardService.delete(id);
 
         return "redirect:/";
